phone-chatbot/daily-pstn-warm-transfer/pyproject.toml
[project]
name = "daily-pstn-warm-transfer"
version = "0.1.0"
description = "Daily PSTN Warm Transfer example with hold music"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[daily,deepgram,cartesia,openai,silero,local-smart-turn-v3,runner,soundfile]>=0.0.101",
    "pipecat-ai-cli",
    "loguru",
    "uvicorn",
    "aiohttp",
    "fastapi",
    "python-dotenv>=1.0.1,<2.0.0",
]

[dependency-groups]
dev = [
    "pre-commit~=4.2.0",
    "ruff~=0.12.1"
]

[tool.ruff]
exclude = [".git", "*_pb2.py"]
line-length = 100

[tool.ruff.lint]
select = ["I"]
ignore = []
